+++ SpeedyWeatherInternals/src/KernelLaunching/KernelLaunching.jl
@@ -25,6 +25,13 @@ function heuristic_workgroup(Wx, Wy, Wz)
     return (min(Wx, cld(cld(256, minWy), minWz)), minWy, minWz)
 end
 
+function heuristic_workgroup(Wx, Wy, Wz, Ww)
+    minWw = min(Ww, 4)
+    minWz = min(Wz, 4)
+    minWy = min(Wy, 4)
+    return (min(Wx, cld(cld(cld(256, minWy), minWz), minWw)), minWy, minWz, minWw)
+end
+
 # WORK ORDER TYPES FOR TYPE-BASED DISPATCH
 """
 Abstract type for different work order patterns in kernel launching.
@@ -52,7 +59,7 @@ Work order for kernels over the diagonal of a (m+1)×m or m×m LowerTriangularAr
 struct DiagonalWorkOrder <: AbstractWorkOrder end
 
 """
-Work order for kernels over a regular 3D array.
+Work order for kernels over a regular N-D array.
 """
 struct ArrayWorkOrder <: AbstractWorkOrder end
 
@@ -100,7 +107,7 @@ end
 
 """
 $(TYPEDSIGNATURES)
-Returns the `workgroup` and `worksize` for launching a kernel over a regular 3D array.
+Returns the `workgroup` and `worksize` for launching a kernel over a regular N-D array.
 """
 function work_layout(::Type{ArrayWorkOrder}, worksize::NTuple{N, Int}) where {N}
     return heuristic_workgroup(worksize...), worksize